sprint 2

+++ src/main/java/no/ntnu/stud/idatt2106/backend/model/base/ExtraResidentType.java
@@ -4,18 +4,18 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-
 /**
  * Entity representing an extra resident type.
  * 
- * <p>Each type specifies the resource consumption (water and food) per resident.
+ * <p>
+ * Each type specifies the resource consumption (water and food) per resident.
  * </p>
  */
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class ExtraResidentType {
-  private int id;
+  private Long id;
   private String name;
   private float consumptionWater;
   private float consumptionFood;
